requirements.txt
fastapi
uvicorn
supabase
python-jose==3.3.0